refactor: pull StatsLeaderboard into its own component
